refactor: auto-load all handlebar partials
