+++ src/templates/builder.ts
@@ -1,29 +1,10 @@
-import fs from 'fs';
-import Handlebars, { compile } from 'handlebars';
+import { compile } from 'handlebars';
 import juice from 'juice';
 import sass from 'sass';
-import { unsubscribe as signUnsubscribe } from '../sign';
+import { unsubscribeLink, loadPartials } from './utils';
 import templates from './';
 
-Handlebars.registerPartial(
-  'layout',
-  fs.readFileSync('./src/templates/partials/layout.hbs', 'utf-8')
-);
-Handlebars.registerPartial(
-  'proposalsHtml',
-  fs.readFileSync('./src/templates/partials/proposals-html.hbs', 'utf-8')
-);
-Handlebars.registerPartial(
-  'proposalsText',
-  fs.readFileSync('./src/templates/partials/proposals-text.hbs', 'utf-8')
-);
-
-export async function unsubscribeLink(email: string) {
-  return `${process.env.FRONT_HOST}/#/unsubscribe?${new URLSearchParams({
-    signature: await signUnsubscribe(email),
-    email: email
-  }).toString()}`;
-}
+loadPartials();
 
 export default async function buildMessage(id: string, params: any) {
   const template = templates[id];
