+++ src/templates/partials/proposals.html.hbs
