+++ src/templates/partials/proposals.text.hbs
